feat: add `LadduError` struct and work in proper error forwarding for reading data and registering `Amplitude`s
`Amplitude`s can no longer be registered with the same name by the same manager. This was causing some issues and there are better ways to handle it.

This also addresses denehoffman/laddu#4. While it doesn't take a Python path directly, it will expand tildes and dots in the input string (as well as environment variables!). Alternatively, users can pass a python `Path` object as a string.