feat: proof-of-concept for Likelihood terms
`LikelihoodTerm`s can be registered in a `LikelihoodManager` just like `Amplitude`s get registered in `Manager`s. The registration returns a `LikelihoodID`, which behaves the same way as an `AmplitudeID` and can be added or multiplied with other `LiklihoodExpression`s. These `LikelihoodExpression`s can then be loaded by the `LikelihoodManager` to create a `LikelihoodEvaluator` which can be minimized in the same way as an `NLL`. Since this is an extended implementation, we also keep the minimization method for `NLL`, since not all `LikelihoodTerm`s *should* be minimized.