fix: properly handle summations in NLL
Floating point addition is not associative, and the standard parallel summation was causing non-deterministic results. This should be fixed with the inclusion of the `accurate` crate in this commit, although I've included a temporary hack for the gradients since you cannot sum `DVector`s using `accurate`. This should eventually be fixed, since I think it might cause a slight unneeded bump in memory (although maybe the compiler optimizes this out?).