feat: update type hints with __ropts__ and add magic methods to easily pickle `Status`
The second of these applies nicely if you want to pickle multiple `Status` objects into the same file with other data, for example, a list of binned fit results. It essentially allows pickle's load(s) and dump(s) methods to work directly on `Status` and any structures which contain `Status` (like a dict or list).