refactor: change order of four-vector components and modify operation of `boost`
This is a *breaking change* and effects every application that uses four-vectors. The `boost_along` method was removed because its usage can be confusing when boosting in a negative direction. Negating a four-vector negates the energy component because these are just regular vectors with a thin wrapper. To avoid confusion and potential mistakes, the method was removed entirely. Only the `boost` method should be used.

The behavior of the `boost` method has also changed to match the Python `vector` and ROOT `TLorentzVector` conventions. Boosting a vector like `p.boost(&-p.beta())` will result in a vector with no three-momentum. Previously, this same operation was done without the `-` sign, but this is confusing since it seems to imply that doubling the velocity of a vector moves it to its center-of-momentum frame.

The order of four-vectors was changed such that energy now comes last: `[px, py, pz, e]`. This was done because the `nalgebra::Vector4` struct has methods like `p.xy()` that get the first two components, which previously would have given `[e, px]` while the user might expect `[px, py]`. This behavior is now the default.

Two fortunate results of writing this in Rust and consolidating everything to one crate. First, the test suite made it easy to refactor. Second, most Python users will be unaffected by this breaking change, even though it is breaking, simply because it does not change any operation outside of raw vector math. Since all the amplitudes are coded in Rust and included in the crate, the modified versions will just work in Python, and no Python amplitudes could exist (yet). If anyone implemented an amplitude in Rust, however, it might need modification, but only if it performs vector operations. Many amplitudes use `Variable`s defined in this crate as a shortcut around such math, so they might also be entirely unaffected.