feat: improvements to `Dataset` conversions and opening methods
add `rest_frame_indices` arguments to Python constructors of `Event` and `Dataset` and add methods to create `Dataset`s from a dict of lists, a dict of `numpy` arrays, a `polars`/`pandas` `DataFrame`, and a `pyarrow` `Table`. Also corrects some instances of `eps` which should be `aux` and allows the `open` method to take a `str | Path` natively.