revert: remove weighted_bootstrap
This concept was flawed, I don't know why I expected it to work, it clearly oversamples events with higher weights because that's what I told it to do, but this obviously doesn't make the total weight of the dataset consistent. There might be a way to do this if I assume the number of events in a bootstrap need not match the original, but that requires some more thinking which I will leave out for now.