feat: add `conj` operator to `Amplitude`s and `Expression`s
Fear not, the gradient is well-defined as long as we are mapping R^n -> C (or so I've read). If the argument was purely real, the conjugate would be a no-op, and if it was purely imaginary, it's just the same as the additive inverse. Mixtures between the two are okay as long as we don't take a derivative with respect to the complex z, as far as I know.