feat: switch to new ganesh API
This is a large rewrite of the structure of the fitting interface, and while most workflows are the same, some things have changed, like the outputs of fits and the settings available. ganesh adds new fitting algorithms and this interface allows us to generate all settings in a more streamlined way. That being said, there will be breaking changes, but they'll hopefully be minor. This also adds `StochasticNLL`s which operate on batches of the data.