Stochastic NLLs and `ganesh` update (#87)
* wip: start generating stochastic nll methods

* feat: add stochastic/batch evaluators

currently, there is no actual stochastic gradient or minimization methods, but there will be when I update `ganesh`

* wip: add stochastic methods to LikelihoodExpression

* wip: start generating stochastic nll methods

* feat: add stochastic/batch evaluators

currently, there is no actual stochastic gradient or minimization methods, but there will be when I update `ganesh`

* wip: add stochastic methods to LikelihoodExpression

* feat: add scalar multiplication to Python Vec3/4

* feat: big revision, update to latest ganesh version and redo fitting/mcmc interface entirely

* feat: switch to new ganesh API

This is a large rewrite of the structure of the fitting interface, and while most workflows are the same, some things have changed, like the outputs of fits and the settings available. ganesh adds new fitting algorithms and this interface allows us to generate all settings in a more streamlined way. That being said, there will be breaking changes, but they'll hopefully be minor. This also adds `StochasticNLL`s which operate on batches of the data.

* fix: update examples

* fix: remove pyarrow and fastparquet to allow for Python 3.14 builds

The release cycle is too long on these projects to support them, Python 3.14 is already released and I'm not waiting around.

* style: remove deprecated lints

* chore: remove unused example script

* style: apply style formatting to all python files

* fix: bump versions, bump MSRV, add actual path to ganesh back in

* ci: attempt to force readthedocs to use a later Rust version

* docs: document getter methods directly to avoid doc duplicates