refactor!: remove `Manager`/`Model` entirely
We can instead track these things internally and have `Amplitude`s
directly produce `Expression`s. This cuts down on the amount of code
both in the crate and in usage, and makes code much more readable in the
process. The largest consequence is that operations on `Expression`s
will panic if two `Amplitude`s share the same name rather than raising
an error. This is the case in Python as well, but I think there may be a
way around it there.